Restructure Electron files
It was getting way too messy in the root of /electron, so I organized
scripts into /electron/src, .html into /electron/pages & moved
greenworks into /electron/lib.

Includes a modification to the build script to remove the .package
folder before building the app. This prevents old files from staying in
there.
