+++ electron/lib/greenworks.js
@@ -9,19 +9,19 @@ var greenworks;
 
 if (process.platform == 'darwin') {
   if (process.arch == 'x64')
-    greenworks = require('./lib/greenworks-osx64');
+    greenworks = require('./greenworks-osx64');
   else if (process.arch == 'ia32')
-    greenworks = require('./lib/greenworks-osx32');
+    greenworks = require('./greenworks-osx32');
 } else if (process.platform == 'win32') {
   if (process.arch == 'x64')
-    greenworks = require('./lib/greenworks-win64');
+    greenworks = require('./greenworks-win64');
   else if (process.arch == 'ia32')
-    greenworks = require('./lib/greenworks-win32');
+    greenworks = require('./greenworks-win32');
 } else if (process.platform == 'linux') {
   if (process.arch == 'x64')
-    greenworks = require('./lib/greenworks-linux64');
+    greenworks = require('./greenworks-linux64');
   else if (process.arch == 'ia32')
-    greenworks = require('./lib/greenworks-linux32');
+    greenworks = require('./greenworks-linux32');
 }
 
 function error_process(err, error_callback) {
