+++ electron/pages/export.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8"/>
     <title>Bitburner</title>
-    <link rel="stylesheet" href="main.css" />
+    <link rel="stylesheet" href="../main.css" />
     <style>
       body {
         background-color: black;
