+++ electron/src/main.js
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const { app, dialog, BrowserWindow, ipcMain } = require("electron");
 const log = require("electron-log");
-const greenworks = require("./greenworks");
+const greenworks = require("../lib/greenworks");
 const api = require("./api-server");
 const gameWindow = require("./gameWindow");
 const achievements = require("./achievements");
@@ -79,7 +79,7 @@ function setStopProcessHandler(app, window, enabled) {
     setTimeout(() => {
       // Wait a few milliseconds to prevent a race condition before loading the exit screen
       window.webContents.stop();
-      window.loadFile("exit.html")
+      window.loadFile("pages/exit.html")
     }, 20);
 
     // Wait 200ms, if the promise has not yet resolved, let's crash the process since we're possibly in a stuck scenario
@@ -191,7 +191,7 @@ app.whenReady().then(async () => {
 
   if (process.argv.includes("--export-save")) {
     const window = new BrowserWindow({ show: false });
-    await window.loadFile("export.html", false);
+    await window.loadFile("pages/export.html", false);
     window.show();
     setStopProcessHandler(app, window, true);
     await utils.exportSave(window);
