+++ electron/src/menu.js
